Splitting out fakes from the sorter_logging_decorator tests
