+++ src/bs/testState.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <vector>
+
+struct testState
+{
+    enum functionCall
+    {
+        sort,
+        logStart,
+        logEnd,
+    };
+
+    std::vector<functionCall> functionsCalled;
+};
